Register worker with coordinator
The coordinator needs to learn about any worker node that becomes
available for processing.
Further, the coordinator requires a private communication channel to
directly tell the worker what to do.
This commit implements a first prototype that enables both.

Workers use the known socket to tell the coordinator that they are
available for work. In addition, they provide an unique socket name,
such that the coordinator is able to talk to them privately for work
assignments.

The coordinator maintains a data structure that stores all known
workers. This will be the basis for further scheduling and health check
algorithms, implemented on the coordinator side.