Improve handling of failure and don't do health check in parallel
Whenever a worker's healthcheck fails, we don't need to resubmit any
tasks to the task queue. That is, because our workers don't persist
their files locally but commit them to the shared file system, while
cleaning them up after they are quitting.
The original paper mentions that all map tasks need to be repeated.
However, this is only the case if the workers keep the map tasks on
their local disk, not if they are kept in the shared file system itself.

There are advantages to letting workers keep their map result on the
local disk, one being that no clean up is required, if the worker does
not write the files to a persistent storage.
A disadvantage is that we need to manage more communication in between
workers, as now reducers need to ask map workers for map results.
Additionally, we have to deal with map workers randomly crashing during
the reduce process, which might stall the map-reduce task, if now we
don't re-assign a busy worker to perform the now again open map tasks.